requirements.txt
google-genai==0.1.0
mysql-connector-python==8.2.0
pymongo==4.6.1
pandas==2.1.4
SQLAlchemy==2.0.23 